requirements.txt
absl-py==2.2.2
albumentations==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asgiref==3.7.2
astroid==2.15.6
-e git+ssh://git@github.com/iankonradjohnson/BatchImageProcessor.git@f85806b50ef8a127d551f0cda4318595ace4c17b#egg=batch_image_processor
beautifulsoup4==4.13.4
black==23.7.0
cachetools==5.5.2
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==3.1.1
contourpy==1.3.2
coverage==7.3.0
cryptography==44.0.2
cycler==0.12.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@400a49c1ec11a18dd25aea3910507bc3bcd15794
dill==0.3.7
distro==1.9.0
Django==4.2.4
dotenv==0.9.9
exceptiongroup==1.1.3
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
fvcore==0.1.5.post20221221
google==3.0.0
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-documentai==3.3.0
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
hydra-core==1.3.2
idna==3.4
image==1.5.33
imageio==2.37.0
iniconfig==2.0.0
iopath==0.1.9
isort==5.12.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.8
layoutparser==0.3.4
lazy-object-proxy==1.9.0
lazy_loader==0.4
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mccabe==0.7.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.24.3
omegaconf==2.3.0
openai==1.75.0
opencv-python==4.8.0.76
opencv-python-headless==4.11.0.86
packaging==23.1
pandas==2.2.3
pathspec==0.11.2
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.3.0
portalocker==3.1.1
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pylint==2.17.5
pyparsing==3.2.3
pypdfium2==4.30.1
pytest==7.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
rsa==4.9.1
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.4.4
sympy==1.13.1
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.0.1
threadpoolctl==3.6.0
tifffile==2025.3.30
tomli==2.0.1
tomlkit==0.12.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.66.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.0.4
Werkzeug==3.1.3
wrapt==1.15.0
yacs==0.1.8
